#44 feat: cart page 파일 위치 수정 및 카트 아이템 총 주문 금액 컴포넌트 적용
